+++ src/App.js
@@ -1,19 +1,15 @@
 import React from "react";
-import { Route, Routes } from "react-router-dom";
 import { useDispatch } from "react-redux";
-import { Home, Favorites, Character, SignUp, Logout, SignIn } from "./pages";
+import { Route, Routes } from "react-router-dom";
 import { Header } from "./components";
-import { getLoggedUser } from "./utils";
-import { setUser } from "./features";
+import { init } from "./features";
+import { Character, Favorites, Home, SignIn, SignUp } from "./pages";
 import "./App.css";
 
 function App() {
   const dispatch = useDispatch();
 
-  const authedUser = getLoggedUser();
-  if (authedUser) {
-    dispatch(setUser(authedUser));
-  }
+  dispatch(init())
 
   return (
     <>
@@ -24,7 +20,6 @@ function App() {
           <Route path="/favorites" element={<Favorites />} />
           <Route path="/signup" element={<SignUp />} />
           <Route path="/signin" element={<SignIn />} />
-          <Route path="/logout" element={<Logout />} />
           <Route path="/character/:charId" element={<Character />} />
         </Routes>
       </main>
