--- src/components/AuthFormRenderer/index.js
@@ -1,1 +1,0 @@
-export * from "./AuthFormRenderer";
