+++ src/components/Header/Search/styles/searchStyle.module.css
@@ -5,9 +5,9 @@
 .search {
   padding: 20px 10px;
   width: 100%;
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
+  box-sizing: border-box;
   font-size: 18px;
   text-align: center;
 }
+
 /*# sourceMappingURL=searchStyle.module.css.map */
