+++ src/components/index.js
@@ -2,4 +2,4 @@ export * from "./Card";
 export * from "./Cards";
 export * from "./Header";
 export * from "./CharacterDescription";
-export * from "./AuthFormRenderer";
+export * from "./AuthForm";
