--- src/features/cards/index.js
@@ -1,1 +1,0 @@
-export * from "./cardsSlice";
