+++ src/features/midlewares/localStorage/LSMiddleware.js
@@ -1,7 +1,18 @@
-import { logout, setUser, toggleFavorites } from "../../user";
-import { toggleObjectField } from "../../../utils";
+import { getLoggedUser, toggleObjectField } from "../../../utils";
+import { init, logout, setUser, toggleFavorites } from "../../user";
 
 export const LSMiddleware = (store) => (next) => (action) => {
+  if (logout.match(action)) {
+    localStorage.setItem("authed", "");
+    return next(action);
+  }
+  else if (init.match(action)){
+    const loggedUser = getLoggedUser();
+    if (loggedUser) {
+      store.dispatch(setUser(loggedUser));
+    }
+    return next(action);
+  }
 
   const users = JSON.parse(localStorage.getItem("users")) || {};
 
@@ -24,9 +35,5 @@ export const LSMiddleware = (store) => (next) => (action) => {
     localStorage.setItem("users", JSON.stringify(users));
   }
 
-  else if (logout.match(action)) {
-    localStorage.setItem("authed", "");
-  }
-
-  next(action);
+  return next(action);
 };
