--- src/pages/Logout/Logout.jsx
@@ -1,16 +1,0 @@
-import { useEffect } from "react";
-import { useNavigate } from "react-router";
-import { useDispatch } from "react-redux";
-import { logout } from "../../features";
-
-export function Logout() {
-  const dispatch = useDispatch();
-
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    dispatch(logout());
-    navigate("/");
-  })
-
-}
