+++ src/pages/SignIn/SignIn.jsx
@@ -1,6 +1,6 @@
 import { useSelector } from "react-redux";
 import { Navigate } from "react-router";
-import { AuthFormRenderer } from "../../components";
+import { AuthForm } from "../../components";
 import { selectLoggedIn } from "../../features";
 import { LOGIN_FORM } from "../../utils";
 
@@ -8,6 +8,6 @@ export function SignIn() {
   const isLoggedIn = useSelector(selectLoggedIn);
 
   return (
-    isLoggedIn ? <Navigate to={"/"}/> : <AuthFormRenderer type={LOGIN_FORM}/>
+    isLoggedIn ? <Navigate to={"/"}/> : <AuthForm type={LOGIN_FORM}/>
   )
 }
