+++ src/pages/SignUp/SignUp.jsx
@@ -1,6 +1,6 @@
 import { useSelector } from "react-redux";
 import { Navigate } from "react-router";
-import { AuthFormRenderer } from "../../components";
+import { AuthForm } from "../../components";
 import { selectLoggedIn } from "../../features";
 import { REGISTER_FORM } from "../../utils";
 
@@ -8,6 +8,6 @@ export function SignUp() {
   const isLoggedIn = useSelector(selectLoggedIn);
 
   return (
-    isLoggedIn ? <Navigate to={"/"} /> : <AuthFormRenderer type={REGISTER_FORM}/>
+    isLoggedIn ? <Navigate to={"/"} /> : <AuthForm type={REGISTER_FORM}/>
   )
 }
