+++ src/pages/index.js
@@ -3,4 +3,3 @@ export * from "./Favorites";
 export * from "./Home";
 export * from "./SignIn/SignIn";
 export * from "./SignUp/SignUp";
-export * from "./Logout";
