+++ src/utils/utils.js
@@ -1,8 +1,8 @@
 export function getLoggedUser() {
-  const nameOfAuthedUser = localStorage.authed ? JSON.parse(localStorage.authed) : "";
+  const nameOfAuthedUser = localStorage.getItem("authed") ? JSON.parse(localStorage.getItem("authed")) : "";
 
   if (nameOfAuthedUser) {
-    const users = JSON.parse(localStorage.users);
+    const users = JSON.parse(localStorage.getItem("users"));
     return users[nameOfAuthedUser];
   }
 
@@ -18,14 +18,14 @@ export function toggleObjectField(object, property) {
 }
 
 export function isUsernameExists(username) {
-  const users = localStorage.users ? JSON.parse(localStorage.users) : [];
+  const users = localStorage.getItem("users") ? JSON.parse(localStorage.getItem("users")) : [];
 
   return users.hasOwnProperty(username);
 }
 
 export function authenticate(username, password) {
   if (!isUsernameExists(username)) return false;
-  const user = JSON.parse(localStorage.users)[username];
+  const user = JSON.parse(localStorage.getItem("users"))[username];
 
   if (username === user.username && password === user.password) {
     return user;
